backend/requirements.txt
requests>=2.31.0
fastapi>=0.115.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
supabase>=2.0.0
uvicorn[standard]>=0.32.0
pandas
openpyxl
